# river/vm: refactor decode order