--- a/pkg/river/vm/tag_cache.go
+++ b/pkg/river/vm/tag_cache.go
@@ -3,6 +3,7 @@ package vm
 import (
 	"reflect"
 	"strings"
+	"sync"
 
 	"github.com/grafana/agent/pkg/river/internal/rivertags"
 )
@@ -10,15 +11,15 @@ import (
 // tagsCache caches the river tags for a struct type. This is never cleared,
 // but since most structs will be statically created throughout the lifetime
 // of the process, this will consume a negligible amount of memory.
-var tagsCache = make(map[reflect.Type]*tagInfo)
+var tagsCache sync.Map
 
 func getCachedTagInfo(t reflect.Type) *tagInfo {
 	if t.Kind() != reflect.Struct {
 		panic("getCachedTagInfo called with non-struct type")
 	}
 
-	if entry, ok := tagsCache[t]; ok {
-		return entry
+	if entry, ok := tagsCache.Load(t); ok {
+		return entry.(*tagInfo)
 	}
 
 	tfs := rivertags.Get(t)
@@ -32,7 +33,7 @@ func getCachedTagInfo(t reflect.Type) *tagInfo {
 		ti.TagLookup[fullName] = tf
 	}
 
-	tagsCache[t] = ti
+	tagsCache.Store(t, ti)
 	return ti
 }