Previously, decoding into a struct would iterate through the tags, and decode each tag in order. Blocks would be grouped by name, and then decoded in into the slice in bulk.

To faciliate #2602, where a slice may have multiple elements from blocks of different names, the iteration order must be reversed. Now, we iterate through statements in the AST, and decode them in order specified in the configuration file. This has the added benefit of returning evaluation errors in order of the config file rather than in order of the tags specified by the Go struct.

A subset of errors has been updated to return `diag.Diagnostic`, as part of #1957 (but does not complete #1957).

Benchmark results show a very minor slowdown in performance, but it is likely within acceptable bounds:

```
name                    old time/op    new time/op    delta
Exprs/object_to_map-10    1.10µs ± 1%    1.11µs ± 1%  +0.44%  (p=0.042 n=19+20)

name                    old alloc/op   new alloc/op   delta
Exprs/object_to_map-10    1.45kB ± 0%    1.45kB ± 0%    ~     (all equal)

name                    old allocs/op  new allocs/op  delta
Exprs/object_to_map-10      19.0 ± 0%      19.0 ± 0%    ~     (all equal)
```